requirements.txt
Flask == 1.1.2
flask-cors == 3.0.8
flask-caching == 1.9.0
flask-bcrypt == 0.7.1
flask-jwt-extended == 3.24.1
flask-sqlalchemy == 2.4.4
flask-migrate == 2.5.3
flask-script == 2.0.6
psycopg2-binary == 2.8.5
flask-restful == 0.3.8
flask-marshmallow == 0.13.0